New development mode - can now fetch latest commit instead of release

+++ scripts/upgradeMaster.sh
@@ -32,6 +32,7 @@ else
 fi
 
 BRANCH="$1"
+DEV_IC_MODE="$2"
 DL_URL="https://github.com/MrKaKisen/nafsdm/archive/"
 MY_VERSION_RAW="`cat /home/master-nafsdm/pythondaemon/version.py`"
 LATEST_VERSION=$(curl https://raw.githubusercontent.com/MrKaKisen/nafsdm/$BRANCH/version.txt)
@@ -66,10 +67,14 @@ fi
 
 echo "* Downloading newest version."
 cd /tmp
-wget $DL_URL$LATEST_VERSION.tar.gz -O nafsdm.tar.gz
-tar -zxvf nafsdm.tar.gz
-mv nafsdm-* nafsdm
-rm -rf nafsdm.tar.gz
+if [ "$DEV_IC_MODE" == "True" ];
+  git clone -b development https://github.com/MrKaKisen/nafsdm.git
+else
+  wget $DL_URL$LATEST_VERSION.tar.gz -O nafsdm.tar.gz
+  tar -zxvf nafsdm.tar.gz
+  mv nafsdm-* nafsdm
+  rm -rf nafsdm.tar.gz
+fi
 
 # req dl
 wget -O requirements.txt $REQ_URL
@@ -230,6 +235,13 @@ elif [ "$MY_VERSION" == "1.2.4-stable" ]; then
 else
   echo "* Oops - something that shouldn't happen, happend anyways."
   exit 1
+fi
+
+# dev set version
+if [ "$DEV_IC_MODE" == "y" ]; then
+  cd /tmp/nafsdm
+  COMMIT_HASH=$(git log -n 1 development | sed -n '1p' | cut -c8-14)
+  echo "version = \"$COMMIT_HASH-dev\"" > /home/master-nafsdm/pythondaemon/version.py
 fi
 
 rm -rf /tmp/nafsdm
