+++ setupMaster.sh
@@ -50,7 +50,7 @@ elif [[ "$OPERATINGSYS" == "debian" ]] || [[ "$OPERATINGSYS" == "ubuntu" ]] ; th
   pip install -r requirements.txt
   rm -rf requirements.txt
 else
-  echo "Invalid operating system. Only 'debian', 'ubuntu' and 'centos' supported."
+  echo "* Invalid operating system. Only 'debian', 'ubuntu' and 'centos' supported."
   exit 1
 fi
 
@@ -58,39 +58,64 @@ fi
 echo "* Fetching information about latest version.."
 LATEST_VERSION=$(curl https://raw.githubusercontent.com/MrKaKisen/nafsdm/master/version.txt)
 
-# select version
-echo "* Please select your version. Type in the version number or type 'latest' for latest version."
-echo -n "* Version: "
-read VERSION_USER
+# commit updates
+echo "* Developers only: Would you like to enable incremental commit updates and use development branch only (y/n)?"
+echo "* Warning: This is a developer function, do not use in a production environment."
+read DEV_IC_CONFIRM
 
-if [ "$VERSION_USER" == "latest" ]; then
-  echo -n "* Confirm? (y/n): "
-  read CONFIRM
-  if [ "$CONFIRM" == "y" ]; then
-    DL_VERSION="$LATEST_VERSION"
+if [ "$DEV_IC_CONFIRM" == "y" ]; then
+  if [ "$OPERATINGSYS" == "centos" ]; then
+    yum install git -y
+  elif [[ "$OPERATINGSYS" == "debian" ]] || [[ "$OPERATINGSYS" == "ubuntu" ]] ; then
+    apt-get install git -y
   else
-    echo "* Aborting.."
+    echo "* Invalid operating system."
     exit 1
+
+  cd /tmp
+  git clone -b development https://github.com/MrKaKisen/nafsdm.git
+elif [ "$DEV_IC_CONFIRM" == "n" ]; then
+  echo "* Skipping.."
+  # select version
+  echo "* Please select your version. Type in the version number or type 'latest' for latest version."
+  echo -n "* Version: "
+  read VERSION_USER
+
+  if [ "$VERSION_USER" == "latest" ]; then
+    echo -n "* Confirm? (y/n): "
+    read CONFIRM
+    if [ "$CONFIRM" == "y" ]; then
+      DL_VERSION="$LATEST_VERSION"
+    else
+      echo "* Aborting.."
+      exit 1
+    fi
+  else
+    echo -n "* Confirm? If version doesn't exist, script will fail. (y/n): "
+    read CONFIRM
+    if [ "$CONFIRM" == "y" ]; then
+      DL_VERSION="$VERSION_USER"
+    else
+      echo "* Aborting.."
+      exit 1
+    fi
   fi
 else
-  echo -n "* Confirm? If version doesn't exist, script will fail. (y/n): "
-  read CONFIRM
-  if [ "$CONFIRM" == "y" ]; then
-    DL_VERSION="$VERSION_USER"
-  else
-    echo "* Aborting.."
-    exit 1
-  fi
-fi
+  echo "* Aborting.."
+  exit 1
 
 echo "* Required packages installed!"
 echo "* Downloading nafsdm & installing.."
 
 # download in temp dir
+if [ "$DEV_IC_CONFIRM" == "n" ]; then
+  cd /tmp
+  wget $DL_URL$DL_VERSION.tar.gz -O nafsdm.tar.gz
+  tar -zxvf nafsdm.tar.gz
+  mv nafsdm-* nafsdm
+fi
+
 cd /tmp
-wget $DL_URL$DL_VERSION.tar.gz -O nafsdm.tar.gz
-tar -zxvf nafsdm.tar.gz
-mv nafsdm-* nafsdm
 
 useradd $USER
 # debian and ubuntu doesn't create its home dir automatically, unlike centos
@@ -112,6 +137,14 @@ chmod +x /home/master-nafsdm/webinterface/start.sh
 
 chmod +x /usr/bin/nafsdmctl
 chmod +x /usr/bin/nafsdm-master
+
+# dev set version
+if [ "$DEV_IC_CONFIRM" == "y" ]; then
+  cd /tmp/nafsdm
+  COMMIT_HASH=$(git log -n 1 development | sed -n '1p' | cut -c8-14)
+  echo "version = \"$COMMIT_HASH-dev\"" > /home/master-nafsdm/pythondaemon/version.py
+  echo "True" > /home/master-nafsdm/pythondaemon/dev_ic_mode.txt
+fi
 
 echo "* Installed. Cleanup.."
